internal/cligen: move splitASCII into split.go
Pure move of the name-splitting engine (splitASCII and its
condAtNearestLetters/searchNearest helpers) out of names.go so it sits
next to split_test.go. The casing functions stay in names.go.
